wells are showing on extent change
