+++ app/src/components/map/LayerLoaderHelpers/DataBCRenderFeaturesNearFeature.tsx
@@ -1,7 +1,7 @@
 import buffer from '@turf/buffer';
 import { getDataFromDataBC } from 'components/map/WFSConsumer';
 import React, { useState, useEffect, useContext } from 'react';
-import { GeoJSON, useMap } from 'react-leaflet';
+import { GeoJSON, useMap, useMapEvent } from 'react-leaflet';
 import { Feature, Geometry } from 'geojson';
 import * as turf from '@turf/turf';
 import pointToLineDistance from '@turf/point-to-line-distance';
@@ -12,6 +12,9 @@ import { Capacitor } from '@capacitor/core';
 import { NetworkContext } from 'contexts/NetworkContext';
 import { DatabaseContext2, query, QueryType } from 'contexts/DatabaseContext2';
 import { WellMarker } from './WellMarker';
+import { q } from '../MapContainer';
+import { createPolygonFromBounds } from './LtlngBoundsToPoly';
+import { MapRequestContext } from 'contexts/MapRequestsContext';
 
 interface IRenderKeyFeaturesNearFeature {
   inputGeo: Feature;
@@ -46,32 +49,34 @@ export const RenderKeyFeaturesNearFeature = (props: IRenderKeyFeaturesNearFeatur
     let nearestWellIndex = null;
     let minDistanceKm = null;
     let wellInside = false;
-    arrayOfWells.forEach((well) => {
-      const turfPolygon = polygon((props.inputGeo.geometry as any).coordinates);
-      const distanceKm = pointToLineDistance(well, polygonToLine(turfPolygon));
-      //label points that are inside the polygon
-      if (turf.inside(well, turfPolygon)) {
-        arrayOfWells[index] = { ...arrayOfWells[index], inside: true, closest: false };
-        wellInside = true;
-      }
-      //set index of the closest well yet
-      if (!!!minDistanceKm || minDistanceKm > distanceKm) {
-        minDistanceKm = distanceKm;
-        if (!arrayOfWells[index].inside) {
-          nearestWellIndex = index;
+    if (turf.booleanWithin(props.inputGeo as any, createPolygonFromBounds(map.getBounds(), map).toGeoJSON())) {
+      arrayOfWells.forEach((well) => {
+        const turfPolygon = polygon((props.inputGeo.geometry as any).coordinates);
+        const distanceKm = pointToLineDistance(well, polygonToLine(turfPolygon));
+        //label points that are inside the polygon
+        if (turf.inside(well, turfPolygon)) {
+          arrayOfWells[index] = { ...arrayOfWells[index], inside: true, closest: false };
+          wellInside = true;
         }
-      }
-      index++;
-    });
-    //label closest well
-    arrayOfWells[nearestWellIndex] = { ...arrayOfWells[nearestWellIndex], closest: true };
-    //set new data to send to ActivityPage
-    if (arrayOfWells[nearestWellIndex].properties) {
-      setWellIdandProximity({
-        id: arrayOfWells[nearestWellIndex].properties.GW_WW_SYSID.toString(),
-        proximity: minDistanceKm * 1000,
-        wellInside: wellInside
+        //set index of the closest well yet
+        if (!!!minDistanceKm || minDistanceKm > distanceKm) {
+          minDistanceKm = distanceKm;
+          if (!arrayOfWells[index].inside) {
+            nearestWellIndex = index;
+          }
+        }
+        index++;
       });
+      //label closest well
+      arrayOfWells[nearestWellIndex] = { ...arrayOfWells[nearestWellIndex], closest: true };
+      //set new data to send to ActivityPage
+      if (arrayOfWells[nearestWellIndex].properties) {
+        setWellIdandProximity({
+          id: arrayOfWells[nearestWellIndex].properties.GW_WW_SYSID.toString(),
+          proximity: minDistanceKm * 1000,
+          wellInside: wellInside
+        });
+      }
     }
     return arrayOfWells;
   };
@@ -88,67 +93,117 @@ export const RenderKeyFeaturesNearFeature = (props: IRenderKeyFeaturesNearFeatur
         layer.bindPopup(popupContent);
       };
 
+  const map = useMap();
+  const mapRequestContext = useContext(MapRequestContext);
+  const { layersSelected } = mapRequestContext;
+  const [lastRequestPushed, setLastRequestPushed] = useState(null);
+  //exclusively used for Async Extent
+  const qRemove = (lastRequestPushed: any, newArray: any) => {
+    q.remove((worker: any) => {
+      if (worker.data && lastRequestPushed?.extent) {
+        if (
+          !turf.booleanWithin(worker.data.extent, lastRequestPushed.extent) &&
+          !turf.booleanOverlap(worker.data.extent, lastRequestPushed.extent)
+        ) {
+          console.log('%cThe new extent does not overlap with and not inside of previous extent!', 'color:red');
+          return true;
+        }
+        if (!newArray.includes(worker.data.layer)) {
+          console.log('%cThe worker in a queue no longer needed as the layers have been changed!', 'color:red');
+          return true;
+        }
+      }
+      return false;
+    });
+  };
+
+  useMapEvent('moveend', () => {
+    let newArray = [];
+    layersSelected.forEach((layer: any) => {
+      if (layer.enabled) {
+        newArray.push(layer.id);
+      }
+    });
+
+    qRemove(lastRequestPushed, newArray);
+
+    if (newArray.length > 0) {
+      newArray.forEach((layer) => {
+        q.push({ extent: createPolygonFromBounds(map.getBounds(), map).toGeoJSON(), layer: layer }, getLayerData);
+
+        setLastRequestPushed({ extent: createPolygonFromBounds(map.getBounds(), map).toGeoJSON(), layer: layer });
+      });
+    }
+    map.invalidateSize();
+  });
+
   const getLayerData = async () => {
-    if (props.inputGeo) {
-      const bufferedGeo = buffer(props.inputGeo, props.proximityInMeters / 1000);
-      if (props.dataBCLayerName === 'WHSE_WATER_MANAGEMENT.GW_WATER_WELLS_WRBC_SVW') {
-        if (Capacitor.getPlatform() !== 'web' && !networkContext.connected) {
-          const res = await query(
-            {
-              type: QueryType.RAW_SQL,
-              sql: `SELECT * FROM layer_data WHERE layerName IN ('well');`
-            },
-            databaseContext
-          );
-          let allFeatures = [];
-          res.forEach((row) => {
+    const mapExtent = createPolygonFromBounds(map.getBounds(), map).toGeoJSON();
+    const bufferedGeo = buffer(mapExtent, props.proximityInMeters / 1000);
+    if (props.dataBCLayerName === 'WHSE_WATER_MANAGEMENT.GW_WATER_WELLS_WRBC_SVW') {
+      if (Capacitor.getPlatform() !== 'web' && !networkContext.connected) {
+        const res = await query(
+          {
+            type: QueryType.RAW_SQL,
+            sql: `SELECT * FROM layer_data WHERE layerName IN ('well');`
+          },
+          databaseContext
+        );
+
+        let allFeatures = [];
+
+        res.forEach((row) => {
+          if (props.inputGeo) {
             const featureArea = JSON.parse(row.featureArea).geometry;
             const featuresInArea = JSON.parse(row.featuresInArea);
 
             if (turf.booleanContains(props.inputGeo, featureArea) || turf.booleanOverlap(props.inputGeo, featureArea)) {
               allFeatures = allFeatures.concat(featuresInArea);
             }
-          });
-          setWellsWithClosest(getClosestWellToPolygon(allFeatures));
-          setKeyval(Math.random()); //NOSONAR
-        } else {
-          getDataFromDataBC(props.dataBCLayerName, bufferedGeo).then((returnVal) => {
-            setWellsWithClosest(getClosestWellToPolygon(returnVal));
-            setKeyval(Math.random()); //NOSONAR
-          }, []);
-        }
-      } else {
-        if (!geosToRender) {
-          if (Capacitor.getPlatform() !== 'web' && !networkContext.connected) {
-            const res = await query(
-              {
-                type: QueryType.RAW_SQL,
-                sql: `SELECT * FROM layer_data WHERE layerName NOT IN ('well');`
-              },
-              databaseContext
-            );
-            let allFeatures = [];
-            res.forEach((row) => {
-              const featuresInArea = JSON.parse(row.featuresInArea);
-              allFeatures = allFeatures.concat(featuresInArea);
-            });
-            setGeosToRender(allFeatures);
           } else {
-            getDataFromDataBC(props.dataBCLayerName, bufferedGeo).then((returnVal) => {
-              console.log(JSON.stringify(returnVal));
-              setGeosToRender(returnVal);
-              setKeyval(Math.random()); //NOSONAR
-            });
+            const featureArea = JSON.parse(row.featureArea).geometry;
+            const featuresInArea = JSON.parse(row.featuresInArea);
+
+            allFeatures = allFeatures.concat(featuresInArea);
           }
-        }
+        });
+
+        setWellsWithClosest(getClosestWellToPolygon(allFeatures));
+        setKeyval(Math.random()); //NOSONAR
+      } else {
+        getDataFromDataBC(props.dataBCLayerName, bufferedGeo).then((returnVal) => {
+          setWellsWithClosest(getClosestWellToPolygon(returnVal));
+          setKeyval(Math.random()); //NOSONAR
+        }, []);
+      }
+    } else {
+      if (Capacitor.getPlatform() !== 'web' && !networkContext.connected) {
+        const res = await query(
+          {
+            type: QueryType.RAW_SQL,
+            sql: `SELECT * FROM layer_data WHERE layerName NOT IN ('well');`
+          },
+          databaseContext
+        );
+        let allFeatures = [];
+        res.forEach((row) => {
+          const featuresInArea = JSON.parse(row.featuresInArea);
+          allFeatures = allFeatures.concat(featuresInArea);
+        });
+        setGeosToRender(allFeatures);
+      } else {
+        getDataFromDataBC(props.dataBCLayerName, bufferedGeo).then((returnVal) => {
+          setGeosToRender(returnVal);
+          setKeyval(Math.random()); //NOSONAR
+        });
       }
     }
   };
 
-  //when new geos received, get well data and run labeling function
-  useEffect(() => {
-    getLayerData();
-  }, [props.inputGeo]);
+  // //when new geos received, get well data and run labeling function
+  // useEffect(() => {
+  //   q.push({ name: 'new extent' }, getLayerData);
+  // }, [props.inputGeo]);
 
   return (
     <>
